+++ app/src/main/java/com/petfolio/infinitus/responsepojo/FetchProductByIdResponse.java
@@ -78,7 +78,7 @@ public class FetchProductByIdResponse {
 
         private String _id;
         private String product_title;
-        private int product_price;
+        private double product_price;
         private int product_discount;
         private CatIdBean cat_id;
         private String threshould;
@@ -114,11 +114,11 @@ public class FetchProductByIdResponse {
         }
 
 
-        public int getProduct_price() {
+        public double getProduct_price() {
             return product_price;
         }
 
-        public void setProduct_price(int product_price) {
+        public void setProduct_price(double product_price) {
             this.product_price = product_price;
 
         }
@@ -449,7 +449,7 @@ public class FetchProductByIdResponse {
             private String _id;
             private String product_img;
             private String product_title;
-            private int product_price;
+            private double product_price;
             private int product_discount;
             private boolean product_fav;
             private double product_rating;
@@ -479,11 +479,11 @@ public class FetchProductByIdResponse {
                 this.product_title = product_title;
             }
 
-            public int getProduct_price() {
+            public double getProduct_price() {
                 return product_price;
             }
 
-            public void setProduct_price(int product_price) {
+            public void setProduct_price(double product_price) {
                 this.product_price = product_price;
             }
