+++ app/src/main/java/com/petfolio/infinitus/responsepojo/VendorFetchOrderDetailsResponse.java
@@ -88,7 +88,7 @@ public class VendorFetchOrderDetailsResponse {
         private String order_id;
         private String product_name;
         private int product_quantity;
-        private int product_price;
+        private double product_price;
         private String prodcut_image;
         private String date_of_booking;
         private String status;
@@ -155,11 +155,11 @@ public class VendorFetchOrderDetailsResponse {
             this.product_quantity = product_quantity;
         }
 
-        public int getProduct_price() {
+        public double getProduct_price() {
             return product_price;
         }
 
-        public void setProduct_price(int product_price) {
+        public void setProduct_price(double product_price) {
             this.product_price = product_price;
         }
